requirements.txt
Django==4.1
pycryptodome==3.15.0
djangorestframework==3.14.0
django-cors-headers==3.13.0
requests==2.28.1
django-filter==22.1
psycopg2==2.9.4
django-jazzmin==2.6.0
coverage==6.5.0
jsonnet==0.18.0
django-nose==1.4.6
django-rest-swagger==2.2.0
selenium==4.7.2
python-gettext==5.0
django-parler==2.3
django-translations==1.3.0
pynose==1.4.8
django-import-export==3.3.4
django-crispy-forms==1.12.0
pillow==10.1.0